+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 1.4.0
+
+* **Breaking change:** Renamed `ScopedContext.get<T>` to `read<T>`.
+* **Breaking change:** Renamed `ScopedListenableFactory` to `ScopedFactory`.
+* Improved error message for `ScopedListenableNotFoundError`.
+* Simplified `ScopedContainer` implementation.
+* Other minor refactorings.
+
+## 1.3.0
+
+* Added `ScopedListenable.builder` for creating `ScopedListenable` widgets with a `BuildContext`.
+
+## 1.2.0
+
+* Added `scoped` and `scope` extension methods.
+
+## 1.1.0
+
+* Renamed `ScopedError` to `ScopedListenableNotFoundError`.
+
 ## 1.0.0
 
 * Initial release.